+++ wazo_agentd/plugins/agent/plugin.py
@@ -1,10 +1,12 @@
-# Copyright 2024 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2024-2025 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from .http import (
     AddAgentToQueue,
     AgentById,
     AgentByNumber,
+    AgentQueuesById,
+    AgentQueuesByNumber,
     LoginAgentById,
     LoginAgentByNumber,
     LoginUserAgent,
@@ -17,6 +19,7 @@ from .http import (
     UnpauseAgentByNumber,
     UnpauseUserAgent,
     UserAgent,
+    UserQueues,
 )
 
 
@@ -40,6 +43,24 @@ class Plugin:
         api.add_resource(
             UserAgent,
             '/users/me/agents',
+            resource_class_args=[service_proxy],
+        )
+
+        api.add_resource(
+            UserQueues,
+            '/users/me/agents/queues',
+            resource_class_args=[service_proxy],
+        )
+
+        api.add_resource(
+            AgentQueuesById,
+            '/agents/by-id/<int:agent_id>/queues',
+            resource_class_args=[service_proxy],
+        )
+
+        api.add_resource(
+            AgentQueuesByNumber,
+            '/agents/by-number/<agent_number>/queues',
             resource_class_args=[service_proxy],
         )