+++ wazo_agentd/plugins/agent/schemas.py
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from xivo.mallow import fields, validate
-from xivo.mallow_helpers import Schema
+from xivo.mallow_helpers import ListSchema, Schema
 
 
 class AgentLoginSchema(Schema):
@@ -24,7 +24,14 @@ class QueueSchema(Schema):
     queue_id = fields.Integer(required=True)
 
 
+class QueueListSchema(ListSchema):
+    default_sort_column = 'id'
+    sort_columns = ['id', 'name', 'display_name']
+    default_direction = 'asc'
+
+
 agent_login_schema = AgentLoginSchema()
 pause_schema = PauseSchema()
 queue_schema = QueueSchema()
 user_agent_login_schema = UserAgentLoginSchema()
+queue_list_schema = QueueListSchema()